// web/src/gen/model/createUserCommand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateUserCommand {
  username: string;
  email: string;
  firstName: string;
  lastName: string;
  roleId: string;
}